pyproject.toml
[project]
name = "youtube-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "resend>=2.11.0",
    "schedule>=1.2.2",
    "yt-dlp>=2025.6.30",
]
